review fix

+++ sql/init.sql
@@ -13,7 +13,7 @@ create table currencies
 (
     id    serial primary key,
     slug  text  not null,
-    value float not null,
+    value decimal not null,
     date  date  not null,
     base  text  not null
 );
